+++ lib/components/DropDown/DropDown.d.ts
@@ -3,7 +3,6 @@ import { IDropDownMenuProps } from './DropDownMenu';
 interface IDropDownProps {
     toggle?: ReactElement;
     menu?: ReactElement<IDropDownMenuProps>;
-    header?: ReactElement;
     onToggleClick?: (e: Event) => void;
 }
 export declare const DropDown: ({ toggle, menu, onToggleClick }: IDropDownProps) => JSX.Element;