+++ lib/components/Snackbar/snackbar.service.d.ts
@@ -3,6 +3,7 @@ export interface ISnackbarService {
 }
 declare class SnackbarService implements ISnackbarService {
     private container;
+    private handler;
     show(message: string, timeout?: number): Promise<void>;
     private hide;
 }